Tracefilters: Computing the fate of flows with respect to individual filters (ACLs) (#1197)
* partial commmit to change context

* one working version

* cleaning pass

* checkstyle

* cleanup

* review comments
