--- projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/questions/ITracerouteQuestion.java
@@ -1,14 +1,0 @@
-package org.batfish.datamodel.questions;
-
-import org.batfish.datamodel.Protocol;
-
-public interface ITracerouteQuestion extends IQuestion {
-
-  void setDst(String dst);
-
-  void setDstProtocol(Protocol protocol);
-
-  void setIngressNode(String ingressNode);
-
-  void setIngressVrf(String ingressVrf);
-}
