+++ projects/question/src/main/java/org/batfish/question/tracefilters/TraceFiltersQuestionPlugin.java
@@ -1,0 +1,22 @@
+package org.batfish.question.tracefilters;
+
+import com.google.auto.service.AutoService;
+import org.batfish.common.Answerer;
+import org.batfish.common.plugin.IBatfish;
+import org.batfish.common.plugin.Plugin;
+import org.batfish.datamodel.questions.Question;
+import org.batfish.question.QuestionPlugin;
+
+@AutoService(Plugin.class)
+public class TraceFiltersQuestionPlugin extends QuestionPlugin {
+
+  @Override
+  protected Answerer createAnswerer(Question question, IBatfish batfish) {
+    return new TraceFiltersAnswerer(question, batfish);
+  }
+
+  @Override
+  protected Question createQuestion() {
+    return new TraceFiltersQuestion(null, null);
+  }
+}
