+++ tests/basic/commands
@@ -36,6 +36,7 @@ test tests/basic/selfAdjacencies.ref get selfAdjacencies
 test tests/basic/traceroute-1-2.ref get traceroute ingressNode="as1core1", dst="host1"
 test tests/basic/traceroute-1-2-ignoreAcls.ref get traceroute ingressNode="as1core1", dst="host1", ignoreAcls=true
 test tests/basic/traceroute-2-1.ref get traceroute ingressNode="host2", dst="1.0.1.1"
+test tests/basic/tracefilters.ref get tracefilters dst="1.1.1.1", nodeRegex="host.*", filterRegex="filter.*"
 test tests/basic/multipath-host1.ref get reachability type="multipath", ingressNodeRegex="host1", srcIps=["2.128.0.0"], dstIps=["3.0.1.2"], ipProtocols=["TCP"], srcPorts=[0], dstPorts=[0]
 test tests/basic/multipath-host2.ref get reachability type="multipath", ingressNodeRegex="host2", srcIps=["2.128.0.0"], dstIps=["1.0.1.1"], ipProtocols=["UDP"], srcPorts=[0], dstPorts=[0]
 test tests/basic/bgpAsnUse.ref get bgpasnuse minCount=2
